add caches

+++ apps/games-api/src/services/session.ts
@@ -8,6 +8,7 @@ import { FastifyRequest } from 'fastify'
 import jwt, { TokenExpiredError, verify } from 'jsonwebtoken'
 import { prisma } from '../shared/db'
 import { env } from '../shared/env'
+import { cache } from '../shared/redis'
 
 enum SessionState {
   Empty,
@@ -20,6 +21,10 @@ export type Session =
   | { state: SessionState.Expired; user: null; token: string }
   | { state: SessionState.Empty; user: null; token?: string }
 
+const sessionCache = cache.entity<string, Session>({
+  keygen: (token: string) => `session:${token}`,
+})
+
 export const getSession = async (req: FastifyRequest): Promise<Session> => {
   if (!req.headers.cookie) {
     return { state: SessionState.Empty, user: null }
@@ -31,6 +36,12 @@ export const getSession = async (req: FastifyRequest): Promise<Session> => {
     return { state: SessionState.Empty, user: null }
   }
 
+  const cached = await sessionCache.get(session)
+
+  if (cached) {
+    return cached
+  }
+
   try {
     verify(session, env.jwt.secret)
   } catch (error) {
@@ -39,7 +50,11 @@ export const getSession = async (req: FastifyRequest): Promise<Session> => {
         ? SessionState.Expired
         : SessionState.Empty
 
-    return { state, user: null, token: session }
+    return await sessionCache.set(session, {
+      state,
+      user: null,
+      token: session,
+    })
   }
 
   const user = await prisma.user.findFirst({
@@ -51,10 +66,17 @@ export const getSession = async (req: FastifyRequest): Promise<Session> => {
   })
 
   if (!user) {
-    return { state: SessionState.Empty, user: null }
+    return await sessionCache.set(session, {
+      state: SessionState.Empty,
+      user: null,
+    })
   }
 
-  return { state: SessionState.Authenticated, user, token: session }
+  return await sessionCache.set(session, {
+    state: SessionState.Authenticated,
+    user,
+    token: session,
+  })
 }
 
 export const getUser = (session: Session): User => {
@@ -77,9 +99,25 @@ async function createSession(options: AddSessionOptions) {
   const expiresAt = new Date(Date.now() + 1000 * expiresIn)
   const token = jwt.sign({ userId }, env.jwt.secret, { expiresIn })
 
-  const session = await prisma.session.create({
+  const user = await prisma.user.findUnique({
+    where: { id: userId },
+  })
+
+  if (!user) {
+    throw new NotAuthenticatedException()
+  }
+
+  const sessionEntity = await prisma.session.create({
     data: { userId, token, expiresAt },
   })
+
+  const session: Session = {
+    state: SessionState.Authenticated,
+    user,
+    token,
+  }
+
+  await sessionCache.set(token, session)
 
   const extraSessions = await prisma.session.findMany({
     where: { userId },
@@ -99,15 +137,15 @@ async function createSession(options: AddSessionOptions) {
     serialize('session', token, {
       domain: env.domain,
       path: '/',
-      expires: session.expiresAt,
+      expires: sessionEntity.expiresAt,
       httpOnly: true,
       sameSite: 'lax',
       secure: true,
     }),
-    serialize('sessionExpiresAt', session.expiresAt.toISOString(), {
+    serialize('sessionExpiresAt', sessionEntity.expiresAt.toISOString(), {
       domain: env.domain,
       path: '/',
-      expires: session.expiresAt,
+      expires: sessionEntity.expiresAt,
       sameSite: 'lax',
       secure: true,
     }),
@@ -120,16 +158,20 @@ async function createSession(options: AddSessionOptions) {
     }),
   ]
 
-  return { cookie }
+  return { cookie, session }
 }
 
 async function removeSession(session: Session) {
-  try {
-    await prisma.session.delete({
-      where: { token: session.token },
-    })
-  } catch {
-    // Session doesn't exist
+  if (session.token) {
+    try {
+      await prisma.session.delete({
+        where: { token: session.token },
+      })
+    } catch {
+      // Session doesn't exist
+    }
+
+    await sessionCache.del(session.token)
   }
 
   const cookie = [
